Add .proto files and generate Go Protobufs
This adds .proto file that matches the current spec at
https://github.com/open-telemetry/opamp-spec

I also added a makefile with gen-proto target to generate
the go Protobufs.

I suggest that we keep the .proto and generated files in
this repo for now to keep initial development fast. Kept in
"internal" for now to avoid leaking unwanted dependency.

We can move the .proto to a separate repository later if we
decide that's the right approach (similar to OTLP proto files).