[DON'T MERGE] Draft client interfaces and types
This is a possible way to dsign the client API. I would like
some feedback on this or possibly an alternate API suggestion
for the client.

This one uses callbacks. I considered a channel-based approach
to notify the agent about messages but it async nature of channels
makes it difficult to return responses to the server.

Any other thoughts on a better API for the client?