[WIP] Draft for supervisor interface and configuration file
This changeset takes a first stab at the supervisor model for
implementing the client side of the OpAMP protocol.

There are two parts to this changeset:

1. Proposal on the supervisor configuration file based on the draft
specification. Two alternate configuration files are provided as an
example of how the supervisor might be instantiated.

2. A simple supervisor interface that all implementations should adhere
to. This allows for the implementation of a generic supervisor as well
as tailor-made ones that has an understanding of the inner workings of
the agent that it is supervising. The supervisor internally wraps the
OpAMP client that communicates with the OpAMP server for management
instructions.