Move messages in StatusReport to AgentToServer message
StatusReport was an unnecessary message, remaining from the time
when we had DataToAgent which had a oneof field, and it was
impossible to set multiple fields that together form the status.
All its fields are moved to AgentToServer message now and can be
set directly there (there is no longer a oneof limitation so now
this is possible to do).

Implements spec change: https://github.com/open-telemetry/opamp-spec/pull/85