Change bitfields from enum to uint64 (#132)
Implements spec change https://github.com/open-telemetry/opamp-spec/pull/125

In some languages enum fields are strongly typed and it is impossible to assigned OR-ed values to the enum field. This makes impossible to compose the bit fields properly.

This changes all bit field declarations from enum to uint64. The enum declarations stay, so that bit definitions are clear.