Remove OnOpampConnectionSettingsAccepted callback
Contributes to https://github.com/open-telemetry/opamp-go/issues/261

I intentionally left undefined the blocking vs nonblocking
requirement for the callback. I suggest that we refine this
after we make implementations and settle on a particular
behavior.

A continuation of this PR is needed that implements steps 4 and 5
of the flow described here:
https://github.com/open-telemetry/opamp-go/issues/261