Fix test TestOfferUpdatedVersion
The test sometimes fails on my machine. The EventuallyExpect function
needs a condition that returns false instead of failing when the
expected message is not received, since a different message may be
sent by the client.

This fixes the test. I run it 100 times and it passes:
`go test -run TestOfferUpdatedVersion/ws -v -race -count 100`