Add support for automatically send connection settings status
Add support to automatically send connection settings status from tha
client to the server. When a client recieves a message with offered
connection settings it will immiedatly send a message stating that it is
applying the settings. If the offered connection settings contain OpAMP
settings the success/failure message will automatically be communicted.
For other connection types the developer will need to call
SetConnectionStatus within the OnMessage callback hook.