Add DownloadHTTPClient callback (#405)
Package Syncer uses the http.DefaultClient and gives users not chance of
configuring an HTTP client for the purpose of downloading packages.
This means users cannot configure TLS, timeouts, keep-alive, use a
custom transport, a proxy, etc to faciliate package file downloads.

This commit add a new DownloadHTTPClient callback that allows users
to provide a custom HTTP Client to use with package downloads.

The callback will be called for every DownloadableFile. This is to
allow users to tweak the client based on the file URL. Users who
do not need to tweak the client for each download can always
return the same client.