Make it possible to set capabilities after client start (#366)
Closes #365 

Corresponds to spec change https://github.com/open-telemetry/opamp-spec/pull/217

This PR introduces a new method on the client – `SetCapabilities`. This PR deprecates the former method of setting `Capabilities` via `StartSettings` in favor of calling `SetCapabilities` prior to start. In the future, the `Capabilities` field will be removed from `StartSettings` and `Start` will error if `SetCapabilities` was not called prior to `Start`. `SetCapabilities` will always ensure that at least `ReportsStatus` is present. Consumers of this method will now be able to change the Capabilities of the Client on demand. It is expected that servers are able to handle these changes.