fix: Add atomic.Bool to ensure wsConnection.Disconnect() is only called once (#359)
Addresses this [issue](https://github.com/open-telemetry/opamp-go/issues/358) by addding an atomic.Bool so we only close the underlying websocket connection once. This PR also adds the test `TestDisconnectServerWSConnection` to ensure that that error log isn't generated when the connection is closed server-side. I renamed `TestDisconnectWSConnection` to `TestDisconnectClientWSConnection` since it tests closing the connection from the client-side, though the comments in the test implied it tested disconnecting from the server.

Perhaps it would be better to simply ignore the error when we close the connection twice? I'm unsure if there are error conditions where we'd want to be able to try to close the connection again.

Resolves https://github.com/open-telemetry/opamp-go/issues/358