Demo: Feat/connection settings status (#390)
* Upate submodule and protobufs

* Add support for automatically send connection settings status

Add support to automatically send connection settings status from tha
client to the server. When a client recieves a message with offered
connection settings it will immiedatly send a message stating that it is
applying the settings. If the offered connection settings contain OpAMP
settings the success/failure message will automatically be communicted.
For other connection types the developer will need to call
SetConnectionStatus within the OnMessage callback hook.

* Update protobufs, rename ConnectionStatus to ConnectionSettingsStatus

* Add OnConnectionSettings callback

Add a new callback: OnConnectionSettings that an agent can use when a
non-OpAMP connection setting is recieved. This callback will only run if
the agent has the capability to run the corresponding connection type.
This callback will also communicate connectionsettingsstatus back to the
server.

* Remove SetConnectionSettingsStatus from client interface

* Remove extra log lines

* Review feedback

* schedule message instead of immediatly sending it

* fix capabilities checking

* Remove unused error

* Add unit test

* Update to latest spec

* Update opamp-spec to v0.13.0