feat: WebSocket server send error callback (#427)
This PR adds functionality to the loop responsible for reading messages from the WebSocket connection to exit if we fail to send a message over the connection. This would trigger a reconnect from the client. Additionally it adds a new callback, OnMessageResponseError, that can be called in this event. This would follow behavior present when we fail to read from the connection.

Resolves https://github.com/open-telemetry/opamp-go/issues/426