fix: Package Download race condition and close idle connections (#434)
Was running into a couple issues here while working on the supervisor to support upgrading the collector.

Data race
When syncing the package, we spin off a go routine to provide periodic updates on the progress. TLDR this go routine doesn't clean up by the time we update the package state with a successful install. This change adds a wait group so we make sure this routine exits before we continue.

Idle go routines
After resolving the data race, was also running into the following error. This was resolved by calling "CloseIdleConnections" on the http client we get to download the file.

If you have any questions about these changes, let me know!